requirements.txt
langgraph>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
pydantic>=2.0.0
streamlit>=1.35.0
pandas>=2.0.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
